cli: skeleton of fetching entities and printing as TOML
[committing code from several days ago]

Very verbose, will likely refactor this significantly. Will probably
move a lot of methods to an impl on API model structs.

At the time of writing this, thinking that golang might have been a
better choice. This is a lot of lines of code for very little
functionality. The context manager trait parameterization on the codegen
API client is very difficult for me to understand; need a macro just to
generate the type? And the lack of a simple default correct
thing to do for error handling is always frustrating.

Regardless, i'm very excited about this tool.